+++ dev/stan.clj
@@ -3,20 +3,37 @@
             [ch.codesmith.anvil.basis :as ab]
             [hato.client :as hc]))
 
+(defn init-url [repository]
+  (str "https://hub.docker.com/v2/repositories/library/" repository "/tags?page_size=100"))
+
+(defn latest-images-search [repository]
+  (let [step (fn [url]
+               (-> url
+                 (hc/get {:as :json})
+                 :body))
+        tags (into []
+               (comp
+                 cat
+                 (keep :name))
+               (iteration
+                 step
+                 :vf :results
+                 :kf :next
+                 :initk (init-url repository)))]
+    (fn [pattern]
+      (first (filter
+               #(re-find pattern %)
+               tags)))))
+
+(defn search-by-pattern [repository search patterns]
+  (into {}
+    (map (fn [[java-version pattern]]
+           [java-version (str repository ":" (search pattern))]))
+    patterns))
+
 (defn latest-images-by-pattern [repository patterns]
-  (let [url      (str "https://hub.docker.com/v2/repositories/library/" repository "/tags?page_size=1000")
-        response (hc/get url {:as :json})
-        tags     (into []
-                   (keep :name)
-                   (:results (:body response)))
-        search   (fn [pattern]
-                   (first (filter
-                            #(re-find pattern %)
-                            tags)))]
-    (into {}
-      (map (fn [[java-version pattern]]
-             [java-version (str repository ":" (search pattern))]))
-      patterns)))
+  (let [search (latest-images-search repository)]
+    (search-by-pattern repository search patterns)))
 
 (def jdk-temurin-image-pattern
   {:java8  #"^8u.*-jdk-jammy"
@@ -25,21 +42,26 @@
    :java18 #"^18\..*-jdk-jammy"
    :java19 #"^19\..*-jdk-jammy"
    :java20 #"^20\..*-jdk-jammy"
-   :java21 #"^21_.*-jdk-jammy"
+   :java21 #"^21\..*-jdk-jammy"
    })
 
-(defn latest-jdk-eclipse-temurin []
-  (latest-images-by-pattern "eclipse-temurin" jdk-temurin-image-pattern))
+(defn latest-jdk-eclipse-temurin [eclipse-temurin-search]
+  (search-by-pattern "eclipse-temurin" eclipse-temurin-search jdk-temurin-image-pattern))
 
 (def jre-temurin-image-pattern
   {:java8  #"^8u.*-jre-jammy"
    :java11 #"^11\..*-jre-jammy"
    :java17 #"^17\..*-jre-jammy"
-   :java21 #"^21_.*-jre-jammy"
+   :java21 #"^21\..*-jre-jammy"
    })
 
-(defn latest-jre-eclipse-temurin []
-  (latest-images-by-pattern "eclipse-temurin" jre-temurin-image-pattern))
+(defn latest-jre-eclipse-temurin [eclipse-temurin-search]
+  (search-by-pattern "eclipse-temurin" eclipse-temurin-search jre-temurin-image-pattern))
+
+(defn latest-eclipse-temurin []
+  (let [search (latest-images-search "eclipse-temurin")]
+    {:jdk (latest-jdk-eclipse-temurin search)
+     :jre (latest-jre-eclipse-temurin search)}))
 
 (def ubuntu-image-pattern
   {:jammy #"jammy-*"})
@@ -49,8 +71,7 @@
 
 (comment
 
-  (latest-jdk-eclipse-temurin)
-  (latest-jre-eclipse-temurin)
+  (latest-eclipse-temurin)
   (latest-ubuntu)
 
   (apps/make-docker-artifact {:main-namespace 'codesmith.anvil.artifacts
@@ -73,7 +94,7 @@
   (keys basis)
 
   (def images
-    (hc/get "https://hub.docker.com/v2/repositories/library/eclipse-temurin/tags?page_size=1000"
+    (hc/get "https://hub.docker.com/v2/repositories/library/eclipse-temurin/tags?page_size=100"
       {:as :json}))
 
   (:tag_last_pushed (first (:results (:body images))))